Move E2E tests into unit tests, add coverage where needed
